Refs #15953 Code refactoring for enrollment renewal factories
